pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-fleet"
version = "0.1.0"
description = "A secure team bootstrapper for isolated Hermes Agent fleets"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nous Research", email = "hermes@nousresearch.com"},
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
hermes-fleet = "hermes_fleet.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools.package-data]
"hermes_fleet" = ["py.typed"]
